+++ RELEASE-NOTES.md
@@ -1,5 +1,13 @@
 # pgEdge Platform Release Notes #############
 
+## done for 23.132 on 2023-09-29
+  - improvements to ACE diff-tables to compare three node clusters 
+  - add `foslots` back to pg15 & pg16
+  - `readonly` now available for pg15 & pg16
+  - `psql` command is now just two args: "sql command" & db
+  - `install pgedge` also now installs `foslots` and `readonly` for pg15 & pg16
+  - fix 'export REPO' bug found by QA
+
 ## done for 23.131 on 2023-09-28
   - improve `cluster` commands by passing env REPO to remote sessions
   - spock 3.2dev2 with fix for trigger ref leak